(BSR)[PRO] refactor: move notification reducer to TS
