+++ pro/src/store/selectors/notificationSelector.ts
@@ -4,6 +4,7 @@ import { Notification } from 'store/reducers/notificationReducer'
 export const notificationSelector = (state: RootState): Notification | null => {
   return state.notification.notification
 }
+
 export const isStickyBarOpenSelector = (state: RootState): boolean => {
   return state.notification.isStickyBarOpen
 }
